requirements.txt
Flask==3.0.0
google-cloud-storage==2.13.0
